Handle the special case of the FAT12/16 root directory better
Don't treat the FAT12/16 root directory as a cluster sequence at all;
don't call assign_clust() for it and instead open-code the relevant
adjustments to the root directory. It simplifies things, and avoids
forcing the root directory to be a multiple of the cluster size.

In strict mode, enforce at least 512 root directory entries.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>