Somewhat saner handling of the CHS geometry for "floppies"
Define a "floppy" as an image without a partition table no more than
5040K in size: 5040K = 10080 sectors = 2*80*63; if the HS geometry is
forced by the user the cylinder count is also limited to 255.

In those cases generate a default CHS geometry that matches the
default for all the standard floppy formats (1-2 heads * up to 80
cylinders * up to 63 sectors.) In that case use drive number 0x00 and
media type 0xf0.

Signed-off-by: H. Peter Anvin (Intel) <hpa@zytor.com>